Angular Project Set-up for Electron Wrapper
Using an angular project to create a desktop electron application.
The set-up is necessary, I do not understand why the default
application builder will not work and why es-build is used
for building the application, I will just
follow what works in the tutorial - just trying this out now.

Not sure how changing the base of the web page will
affect the overall project or electron, but this is necessary
and works just as the tutorial worked.

Youtube tutorial by Joshua Morony.